feat(charts): groudedTable
