+++ packages/plugin-block-charts/src/client/chart/antd/tools/renderByConfig.ts
@@ -1,0 +1,4 @@
+export function renderByConfig(text, record, configs) {
+  //  TODO: 在渲染阶段先进行数据转换机制, 然后走用户的配置机制; transform
+  return text;
+}
